easi/simulators/ai2thor/v5_0_0/requirements.txt
ai2thor>=5.0.0,<6.0.0
numpy>=1.24
Pillow>=10.0
